IP-38 final schedules/next

+++ busrush/src/main/java/ies/project/busrush/dto/busrush/NextScheduleDto.java
@@ -12,5 +12,5 @@ import java.time.LocalTime;
 public class NextScheduleDto {
     private RouteBasicDto route;
     private LocalTime time;
-    private Integer delay;
+    private Double delay;
 }
